move plugins to modules
