+++ lua/plugins/tokyonight.lua
@@ -1,0 +1,13 @@
+return {
+  'folke/tokyonight.nvim',
+  priority = 1000, -- Make sure to load this before all the other start plugins.
+  config = function()
+    ---@diagnostic disable-next-line: missing-fields
+    require('tokyonight').setup {
+      styles = {
+        comments = { italic = false }, -- Disable italics in comments
+      },
+    }
+    vim.cmd.colorscheme 'tokyonight-night'
+  end,
+}
